User and Post methods for CRUD implemented - functional and running

+++ src/main/java/com/example/repository/FriendshipRepository.java
@@ -3,5 +3,7 @@ package com.example.repository;
 import com.example.entity.Friendship;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FriendshipRepository extends JpaRepository<Friendship, Integer> { }
+
+public interface FriendshipRepository extends JpaRepository<Friendship, Long> {
+}
