+++ src/main/java/com/example/repository/UserRepository.java
@@ -2,7 +2,6 @@ package com.example.repository;
 
 import com.example.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
-public interface UserRepository  extends JpaRepository<User, Integer> {
-
+public interface UserRepository extends JpaRepository<User, Long> {
 
 }
